pyproject.toml
[project]
name = "toolhub"
version = "0.1.0"
description = "Local documentation index for AI coding agents"
readme = "README.md"
authors = [
    { name = "natustx", email = "nathan.little@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "lancedb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "gitpython>=3.1.0",
    "rich>=13.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "tomli-w>=1.0.0",
    "pydantic>=2.0.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.11.0",
]

[project.scripts]
toolhub = "toolhub.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
